refactor: successful refactor - got it to boot

+++ src/include/driver/fb/fb.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <arch/noarch/fb/fb.h>
-#include "constants/color.hpp"
+#include <arch/noarch/fb/constants/color.hpp>
 
 // Draw a single pixel to the framebuffer
-void putpixel(int x, int y, unsigned long color);
+void putpixel(int x, int y, uint32_t color);
